requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-core==2.20.1
orjson==3.10.7
python-multipart==0.0.9
psycopg[binary]==3.2.10

